srtgo-web/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
celery==5.3.4
redis==5.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0
aiofiles==23.2.1
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Original srtgo dependencies
click==8.1.7
curl_cffi==0.6.2
inquirer==3.2.0
keyring==24.3.0
PyCryptodome==3.19.0
prompt_toolkit==3.0.42
python-telegram-bot==20.7
termcolor==2.4.0